Still developing radial profile design.
Included an inverse profile method, which will probably get eliminated.
